webadmin: Move Network QoS logic to common ground
Made the following changes to allow Host Network QoS access to logic
implemented for VM Network QoS:

* Created an AsyncDataProvider method to conveniently retrieve
NetworkQoS entities for a given DC, which also adds the empty QoS
(unlimited) to the top of the list.

* Moved the empty network QoS definition from VnicProfileModel to
NetworkQoSModel. It's the best location I found for it in the
frontend, otherwise it belongs in the common project (but the backend
doesn't use it).

* Moved the QoS selection logic to Linq.

Change-Id: Ib32cfe2392e86431323176fa9ef3e0aa00dbb841
Signed-off-by: Lior Vernia <lvernia@redhat.com>