packaging: spec: fedora: provide our own apache-sshd
Temporary workaround until apache-sshd-0.9.0 which is currently in
fedora will be downgraded to at least apache-sshd-0.8.0 or upstream will
provide fix which will be included in newer version of fedora package.

Another option is to add a conflict, however we do not know what we
conflicting with, so safer is just to provide our own jar for now.

Related-To: https://bugzilla.redhat.com/show_bug.cgi?id=1021273
Change-Id: I277c005969d57adaa5dc4f36944bc47e2103a02f
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>