restapi: Corrected the resolve action for copy mode
Modified to correctly copy the hook content from a provided node to
others in case of resolve from host content option.
Earlier it was always copying the engine version of the hook content to
all the nodes in case a resolve from a specific node was executed.

Usage:
    api/clusters/<cluster-id>/glusterhooks/<hook-id>/resolve

Input Parameters:
    <action>
        <resolution_type>COPY</resolution_type>
        <host id="<Host-Id>"/>
    </action>

    or

    <action>
        <resolution_type>COPY</resolution_type>
        <host>
            <name>Host-Name</name>
        </host>
    </action>

Change-Id: I412e582addc2bcb8d7176634db7eafaa4686668b
Bug-Url: https://bugzilla.redhat.com/1033772
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>