core: remove pending VMs counting
The number of pending VMs is not in use anywhere, but it is still
computed. In order to compute the pending VMs we increment and decrement
the counter while running VMs in synchronized blocks and it is not good
in terms of performance. So since this code is redundant, it is better
to remove it.

Change-Id: I3bd7a757be2979b86227ce56064e6165080035eb
Signed-off-by: Arik Hadas <ahadas@redhat.com>