core: fixing dual sequential PM scenario
In case that dual sequential power management cards are defined PM
operations attempt to use the first card (primary) first and only if
that
failed attempt to use the second card (secondary).
When the primary card is not defined correctly the secondary PM fails to
take place.
This patch fixes this issue by insuring that we have a valid proxy host
prior to the secondary PM card operation.

Change-Id: Id88b22f06241c02f12dbc0efe32eb7f04063ed06
Bug-Url : https://bugzilla.redhat.com/show_bug.cgi?id=1013654
Signed-off-by: Eli Mesika <emesika@redhat.com>