engine: Show waiting for gluster tasks
When a gluster task is run on a volume for
the second time, before gluster returns the
newly created task, the view was fetching the
previous task's status.
Changed the view where condition to fetch the
status correctly.

Also, the view should only return job status
related to gluster async tasks.

Change-Id: Ia3a2ac3ddc2f46f24b84628cf52f3a193cdebf31
Bug-Url: https://bugzilla.redhat.com/1025304
Signed-off-by: Sahina Bose <sabose@redhat.com>