core: Add QoS member to Network and VdsNetworkInterface
Added a NetworkQoS member to these entities. The QoS member of the
Network entity is represented by its ID and refers to the QoS that
should be configured on host interfaces to which the network is
attached, while that of the VdsNetworkInterface entity is implemented
as an integral part of the entity and refers to the QoS reported by
VDSM on a host.

Change-Id: I63776837d41c620258fa53b9a1335b76a5971cea
Signed-off-by: Lior Vernia <lvernia@redhat.com>