engine: Set brick port and pid when returned
When gluster process is restarted, the gluster
CLI returns the brick port as NA. This causes
the VDS xml return handler to fail as the brick
port is expected as an integer value.

Added code to set the int values only if
returned correctly.
If brick port is N/A then brick's status should
be set offline.

Change-Id: I544b0aa4705b4bcde7638f15b5b235ba10deb3b1
Bug-Url: https://bugzilla.redhat.com/1036039
Signed-off-by: Sahina Bose <sabose@redhat.com>