core: allow current master domain to be selected as the new master
Currently the current master domain cannot be selected as the new
master, which can cause to the system to not be able to recover in case
that it's the only available domain.

This patch allows to perform reconstruct using the master domain in case
that we know that it's accessible -
1. Getting WrongMasterDomain error from vdsm means that we could get to
the master domain metadata.

NOTE: in that case, the reconstruct command will be executed with
parameter indicating that the current master can be selected as the new
master - but the host that return the WrongMasterDomain error or the master
domain id as the master domain id as the id of the domain to reconstruct to aren't
being passed to the command - this is being done to avoid possible
issues (for example: failing to reconstruct to that domain although we
can read the metadata, keep trying through the same host..etc).
The intention here is to improve the current situation for most of the
usecases.

2. When the master domain info in the db is different then the
info received from getStoragePoolInfo.

Change-Id: I705298d4f94b4afacc3ca028bb0a4e6080ac5a46
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1016118
Signed-off-by: Liron Aravot <laravot@redhat.com>