webadmin : Retain bricks refresh
On clicking retain brick button from remove-brick status
pop up,after succesfull execution of the command the pop up
was not getting refreshed and hence the stopped at time
was not getting displayed.But,on closing the pop up and
reopening it again,with the new pop up the stopped at time
was getting displayed.
Hence,this patch fixes it by refreshing the pop up after
performing retain-brick option.

Change-Id: Iddc56eec6f9e456c6fdcf6dd45152b58fb0878ed
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1037583
Signed-off-by: Anmol Babu <anbabu@redhat.com>