webadmin: Simplify VnicProfile flow in NetworkModel
The previous implementation was over-complicated, and allowed users to
perform the same actions from both the Edit Network dialog and the
Network/Profiles subtab. It is now only possible to create profiles
for a newly-created network, or create profiles for a network that
used to be a non-VM network and is turned into a VM network (which
from a profiles point of view is the same as a new network).

Change-Id: I21b632e7625e24a6ecb28e278639805990f6af41
Bug-Url: https://bugzilla.redhat.com/1031558
Signed-off-by: Lior Vernia <lvernia@redhat.com>