webadmin: Only carry on with import if network added
When importing a network from a provider, a sequence of backend
commands is triggered; first creation of a network, then attachment to
cluster and in the end creation of a default vNIC profile. Previously
the succeess of the network creation wasn't verified, leading to
confusing errors from the succeeding backend commands. This is fixed
here.

Change-Id: Ie718f92f724f71e6e9d02ad19e4c7912efe5bb94
Bug-Url: https://bugzilla.redhat.com/1031366
Signed-off-by: Lior Vernia <lvernia@redhat.com>