webadmin: Changed semantics of NULL_OPERATION in Setup Networks
Changed the semantics of a generic NULL_OPERATION to actually
represent a NO-OP. When a generic NULL_OPERATION is performed no
message appears in the header of the dialog, thus promoting a feel
that no operation is performed (rather than an operation is attempted
and failed).

Also removed a check whether an interface being dragged is already
bonded; when dragging an interface that's part of a bond it is the
bond that's dragged anyway, so this check can't return true.

Change-Id: Ie5188b3c6ca5838f6e0d6bb184e554970b6f0c0f
Bug-Url: https://bugzilla.redhat.com/906313
Signed-off-by: Lior Vernia <lvernia@redhat.com>