tools: Fixing edit action of manage-domains
Bug-Url: https://bugzilla.redhat.com/1037894

The following patches fixes the edit action of
manage-domains tool to support edit of values provided
by the "ldapServers" argument.
Prior to this fix, the change was not stored at the
proper location at the vdc_options table.
Bare in mind that if -ldapServers is not provided during
the edit action no change will be made to the entry.

Change-Id: I9f86b8ce2c697931b9c8388ea87fe07d557530c6
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>