core: get cd path only when the VM isn't the hosted engine
When the VM we wish to change disk to is the hosted engine VM, getting
the CD path fails, as this VM isn't managed by the engine.
For that reason, this patch moves this logic to after we know it is not
the hosted engine. In addition, it is good to do that not in the
constructor but after some validation is done, as it is less error
prone.

Change-Id: Ie94993f2f43b5c3d2658542f38292b0d8c5503b3
Bug-Url: https://bugzilla.redhat.com/1014030
Signed-off-by: Oved Ourfali <oourfali@redhat.com>