tools: notifier: Refactor model and DB mapping of eventNotificationMethods.
Main change in this patch set is storing eventNotificationMethods
as enum in java objects and as an sql string in the event_subscribers
table(and not int). The  mapping is easier and the db is human readable.

Another side effect of this change is there is no more need for
some mapping classes deleted in this patch. They are all replaced
by the simpler NotificationMethodsMapper. The table
event_notification_methods was dropped since it duplicates the enum
eventNotificationMethods and replaced by new an sql CHECK constraint.

Change-Id: I8b71c4e78bbdca3d02d2ac4ef419b9d3d7d58761
Signed-off-by: Mooli Tayer <mtayer@redhat.com>