webadmin : Hide virt fields in the host->general sub-tab in Gluster mode
Not adding instead of just hiding virt-related fields in the host -> general
sub-tab and hence fix the allignment issues.
       Modified the SubTabHostGeneralView to NOT take row numbers statically,
as doing so,will introduce blank spaces if a particular row is made
invisible in a particular application mode.
So,modified FormBuilder to have a map("nextEmptyRowForColumn") of first unused
rows for each column and hence decide dynamically which row to insert the current
widget into for a given column.

Change-Id: I770547907c681a4648999c9d0d3d3ff7aba93971
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1025101
Signed-off-by: Anmol Babu <anbabu@redhat.com>