core, restapi: Add DbGroup
This change introduces a new DbGroup entity, very similar in purpose to
the existing DbUser entity. Initially this new entity has been mostly
copied from LdapGroup and replaces it in places where LDAP concepts
shouldn't be involved. The existence of this entity will simplify the
separation of concerns between the persistence and authentication
subsystems.

Change-Id: Idb1a7146c29eb74f97e10043d65b5a67f1430021
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>