restapi: Add per-VM migration_downtime to REST API
Added migration_downtime field to VM representing maximum number of
milliseconds for downtime during live migration.

Due to problems representing the Integer value of 'null' which has
semantic meaning on the backend, in the API we represent this "use
default" option using integer "-1" which is outside the range of
valid timeouts.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1055434
Change-Id: I53eaecc790d8805f55417f7eb74095001a325140
Signed-off-by: Martin Betak <mbetak@redhat.com>