core: more accurate scores for ED weight module
Even Distribution (ED) weight module inner calculation score
is a real number, and we expose it as an integer by rounding it
up.
Currently we round it up (Math.ceil), but to be more accurate we should round
it to nearest integer (Math.round).
To avoid 0 adding 1 to the score.

Change-Id: Ib668cf266629eb4e0f2c1cbd5ac893dc7deac19c
Bug-Url: https://bugzilla.redhat.com/1040460
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>