4. core: Use DbGroup to add permissions
Currently the operation that adds permissions receives as parameters a
user, a group and the permission to be added. The user and group
parameters are populated by the client when the user or group needs to
be added to the database before the permission. The type of the user is
DbUser and the type of the group is LdapGroup, which is not very
consistent.
This patch changes the type of the group parameter to DbGroup.

Change-Id: Iae84e0bd42c8171366fa09121f66243f9b98ad2d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>