core,engine,webadmin: Automatic power management of hosts
Add bll and UI support for policy based power management
of hosts.

VdsStatic contains a new flag that can disable the automatic
mechanisms. There is a new chackbox in the Host/Power management
subtab to control this.

VdsDynamic contains a new flag that is used to distinguish
between hosts that were put to maintenance/down automatically
and those that were put there by user.

http://www.ovirt.org/Features/HostPowerManagementPolicy

Change-Id: Icdea8bdef15556be1591ffb7554f79238f865ec8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1035238
Signed-off-by: Martin Sivak <msivak@redhat.com>