core: Mark management network as display and migration
Introducing an upgrade script to mark the management network as
display and/or migration, if no other network in the cluster has that
role. This is implicitly the case anyway because when these roles
haven't been assigned they fall back to the management network, this
makes it explicit.

Also added code to assign the roles to the management network when
another network loses the role, or when a new cluster is created.

Change-Id: I73bf3c583ae28bf32d7c602ec9f4b0fb5c7f391d
Bug-Url: https://bugzilla.redhat.com/1023722
Signed-off-by: Lior Vernia <lvernia@redhat.com>