frontend: ignore in findbugs EventArgs unconfirmed cast errors
Ignore in findbugs errors related to casting EventArgs.
Fix some existing exclude filters that stopped working due to refactoring
introduced in some recent patches that caused the order and the number of anonymous classes that contained findbugs warnings to change.
The casting is safe in this case, and a future refactoring
will make findbug warning resolved in any case.(RFE #1034844)
This is patch #10 following
http://gerrit.ovirt.org/#/c/22230/

Change-Id: I1722507982e53f91dfd1f932f9b463cb5e72c2c7
Signed-off-by: Alissa Bonas <abonas@redhat.com>