webadmin: Fixed buggy profiles in edit network dialog
A recent careless bug fix introduced a couple of bugs to the profiles
section of the edit network dialog, e.g. an attempt to create profiles
is made every time a network is updated and it is impossible to
approve the dialog when the list of profiles is empty. Hopefully this
fixed everything.

Change-Id: I20ed93c4b0a02118ce9732ce7db3c446b8a07803
Bug-Url: https://bugzilla.redhat.com/1043386
Signed-off-by: Lior Vernia <lvernia@redhat.com>