frontend: ignore in findbugs EventArgs unconfirmed cast errors
Ignore in findbugs errors related to casting EventArgs.
Fix existing exclude filters that stopped working due to
refactoring introduced in some recent patches
that caused the order and the number
of anonymous classes that contained findbugs warnings to change.
The casting is safe in this case, and a future refactoring
will make findbug warning resolved in any case.(RFE #1034844)

Change-Id: I60fe6a07a43d4aab804823f7946f17f494339e33
Signed-off-by: Alissa Bonas <abonas@redhat.com>