engine: Update Power Saving policy with power management
Add the logic that controls the automatic power management to
PowerSavingBalancePolicyUnit.

The unit got two new parameters:
- HostsInReserve - the amount of hosts with no VMs that should
                   be kept running
- EnableAutomaticHostPowerManagement - the feature enable flag

When there are too many hosts with no VMs, the algorithm will
select a host and put it first to maintenance and then down.

If we need more hosts, it will start them up.

http://www.ovirt.org/Features/HostPowerManagementPolicy

Change-Id: I2da9f4e019c3dd006caefbaa9baf9a533ce41f3d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1035238
Signed-off-by: Martin Sivak <msivak@redhat.com>