engine: Gluster peer probe and status on same server
Gluster peer status and volume info results are
served from the local db of the node, and hence the
results are not consistent if the commands are executed
on different nodes.
- Server is set to UP after a gluster peer probe only
when it is returned in the peer list from another node.
- Changing the up server selection to return the same
server to avoid inconsistent results for non distributed
gluster commands.

Using getRandomUpServer while calling distributed
gluster commands.

Change-Id: I1359a0c8f270b70a6afd3fd1490e5bb602d9b03f
Bug-Url: https://bugzilla.redhat.com/1043032
Signed-off-by: Sahina Bose <sabose@redhat.com>