core: GetAllDisksByVmIdQuery - improved performance
Currently when executing GetAllDisksByVmIdQuery, all the vm disks are
queried using the GetDisksVmGuid stored procedure and then the vm device
for each disk is being loaded to complete to queried data.

This can cause to mass of loads that could be prevented, GetDisksVmGuid already
performs join with vm_device table so that records could be just
returned.

Change-Id: I91d12ee535563204b5d6aaeb64ec6a7c23ea1a81
Bug-Url: https://bugzilla.redhat.com/1130065
Signed-off-by: Liron Aravot <laravot@redhat.com>