core, engine: sPAPR SCSI on PPC64 VMs
This change adds support for the sPAPR VSCSI on PPC64 guests, which is
a SCSI controller specific to this platform. It also includes code
necessary to create virtual SCSI CD-ROM drives connected to this
controller, and add disks to it. In PPC64 guests there are two SCSI
controllers available, the VirtIO-SCSI and the sPAPR VSCSI. In order to
create disks in both of them, the entire address of each disk must be
set by the engine, because libvirt does not support partially adressing
each disk by setting only the controller selected by the user.

Change-Id: Id4fefa225cb44f0e09a7a056e07b8393dda6819b
Signed-off-by: Vitor de Lima <vitor.lima@eldorado.org.br>