core: setting NIC boot order by NIC name
The default sorting of network interfaces is according to their MAC
addresses. The boot order of network interfaces was also used to be
set according to the order of their MAC addresses.

We're now changing the way we set the boot order for NICs such that it
will be set according to the order of their names instead of the order
of their MAC addresses.

Change-Id: I7916c3873361ef5bc4331ddfd3c60506f3c99fa3
Bug-Url: https://bugzilla.redhat.com/1045022
Signed-off-by: Arik Hadas <ahadas@redhat.com>