packaging: setup: customize NFS export ACL
ACL is always input and no attempt is done to preserve previous values.
Export line is always regenerated. If there is /etc/exports.d, the file
there is always regenerated and contains a suitable comment.

If path is exported in /etc/exports and there is /etc/exports.d, we
delete the line from /etc/exports.

Change-Id: Ia35aaf42cb46247a2a02df1c2418f4613ad6b965
Signed-off-by: Yedidyah Bar David <didi@redhat.com>