engine: Change AddNetworkCommand to non-transactive
The AddNetworkCommand will be used also for configuring
the network on the hosts by its label. For that purpose
the class should be managed as a non-transactive, as
changes should be kept into the database prior to setup
networks and also the network update shouldn't be part
of a wider transaction.

Change-Id: I696601506a76db8fe9d0bef790b9cd7c2b434449
Signed-off-by: Moti Asayag <masayag@redhat.com>