core: Fixing NPE when group is assigned with permissions
The patch associated with Idb1a7146c29eb74f97e10043d65b5a67f1430021
introduced a regression in assigning a group permissions.
Prior to that patch the entity DbGroup did not exist, hence
the CTOR of LdapGroup that has DbGroup did not exist as well.
The memberOf initialization is fixed here for cases that the flows do not
set the memberOf attribute

Change-Id: If92ce2718a32658b4118e4e7f9334b631fa6c839
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>