core: setting all flags to true by default
Gluster had already updated the behaviour of configuration flags to be
true by default and to be set explicitly to false for the versions that
the flag is not supported.
This patch implement this method on all configuration flags and removes
from 0000_config.sql redundant calls.
If a value for a requested version is not found, the default value in
the ConfigValues.java file a annotation is used.

Change-Id: I467ce4e0e483af129559dae87629d2cd4fcf0221
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=996978
Signed-off-by: Eli Mesika <emesika@redhat.com>