core: Reset emulated machines after architecture changing
After PPC patches, cluster received a new parameter, the architecture
one. As it affects the parameter "emulated machine", when a change
occurs on CPU name, which results in an architecture change, the
parameter "emulated machine" must be updated.
This patch introduces the change to set the cluster to reset this
parameter when adding the first host.
Important: It will happen only with empty clusters, since there are
validations when changing CPU name for other architecture.

Change-Id: I1d400b8d475f63e720ad88ad00116aea9720f7ba
Signed-off-by: Leonardo Bianconi <leonardo.bianconi@eldorado.org.br>