core: Findbugs exclussions in JBoss modules plugin
The code generated by Maven for the JBoss modules plugin contains a
class that has an minor issue pointed out by findbugs. These issue has
already been reported to the maven developers:

  http://jira.codehaus.org/browse/MPLUGIN-255

While this is solved in Maven we are adding a findbugs configuration to
ignore it. This configuration is the same used in all the rest of the
engine modules, the only difference is the exclussions list.

Change-Id: I76106b935e0e987db04d99e36f89adc536754021
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>