core: remove redundant fetch of vm disks on run vm
There is no need to fetch the VM disks from the DB after we start to
create stateless images on run VM as stateless flow. We fetch them in
order to update the VM which is local to the RunVmCommand instance with
the stateless images. But it is not used later on in the execute phase
of the command, and the end-action phase is invoked on a new instance of
RunVmCommand anyway.

Change-Id: I39f761ce6ed9563dfc1473afd1e7d330bc95bb20
Signed-off-by: Arik Hadas <ahadas@redhat.com>