core: fix regression that cleared run once flag
We store a flag that indicates whether the VM is run in run once mode or
not. The batch updates in VdsUpdateRuntimeInfo didn't update this flag
correctly, which caused it to be cleared always by the monitoring.

This patch fix this regression by adding the run-once attribute to the
mapper that is used by the batch updates of VmDynamic.

Change-Id: Ifc2bfe9c500f649dff39f4bbfae6d4b7e57cf883
Signed-off-by: Arik Hadas <ahadas@redhat.com>