engine: Update network on hosts when label is changed
When the network's label is being modifies, the change
to the label should be reflected to the labeled hosts.

There supported options in this regards are:
1. Labeling a network - will add the network to any of
labeled hosts which the network is assigned to their
cluster.
2. Unlabeling a network - will remove the network from
any of labeled hosts that the network is assigned to
their cluster.
3. Label unchanged - only sync action will apply to
hosts that the network is configured on.

Changing an existing label while the label is set on
interfaces will not be supported at this time.

Change-Id: Id8b851d589db8ad6ef025d3b752c9596aa30aee2
Bug-Url: https://bugzilla.redhat.com/1040586
Signed-off-by: Moti Asayag <masayag@redhat.com>