core: Introduce NetworkLabel entity
The NetworkLabel entity is a plain container of the
label id. The class designed to be transfered to
the clients when querying for the label information
which is managed either as part of the Network
entity or as part of the VdsNetworkInterface.
Having the label as an explicit entity will standardize
data transfer between the restapi to the engine.

Change-Id: I28f4e7688a02cdc8bf0475dbd237e09352add2e3
Signed-off-by: Moti Asayag <masayag@redhat.com>