webadmin: Update network label GUI field
Moved it from the provider section in the Add/Edit Network dialog, as
now it has significance for an internal network as well. Changed it to
be a suggest box instead of a text box, where in case the network is
internal, the existing labels in the DC are suggested.

Change-Id: Ib0ed563897fd1d7eb4acc9ff322373132f01f8c3
Signed-off-by: Lior Vernia <lvernia@redhat.com>