core: Allow to fence host manually in reboot state
In the event of a full host power outage (including fence devices) we
are waiting for the network attempts to reconnect while the Host is in
"Connecting" state in which manual fence is forbidden since this is
considered as a INTERMITTENT state.
After that, the Host non-responding treatment is taking place and trying
to manually fence the Host during the fencing retries will fail
since manual fence calls ClearNonResponsiveVdsVmsCommand which currently
allows only Hosts that are non-responding to execute this command.
That means that VMs can be freed only after the network reconnect
attempts plus all fencing handling took place which is quite a long time (19 min
according to the BZ description)

The fencing flow moves the Host to "Reboot" state at the time the
Restart command is issued.

This patch modifies the ClearNonResponsiveVdsVmsCommand canDoAction
method to allow execution of this command while Host is in "Reboot"
state in order to shorten the above time (in the flow described in the
BZ this patch allows freeing the VMs at 2013-12-13 12:40:33 instead of
at 12-13 12:52:39 which will save us more than 12 min)

Since when the Host is manually fenced the user approves that the Host
was rebooted we have no problem to allow the user freeing the VMs
earlier if he knows the real state.

Change-Id: I4f5f45ca122d841cc411cc6d9541d0903aeb955a
Bug-Url : https://bugzilla.redhat.com/show_bug.cgi?id=1044091
Signed-off-by: Eli Mesika <emesika@redhat.com>