Fixing a bug in Alert logging.
Log Alert when failing to get Host status via the PM agent was recorded
to the database without the Host UUID, therefor the attempt to remove
this Alert when the Host was fenced manually failed and a junk Alert
remained in the database.

This patch sets the Host id on the Alert log message.

Change-Id: Iee997e6b3df37f411a74815757195dc584c4e8e5
Bug-Url : https://bugzilla.redhat.com/show_bug.cgi?id=1045139
Signed-off-by: Eli Mesika <emesika@redhat.com>