engine: Use an ancestor type in Entities helper class
There is no reason to limit the signature of
Entities.hostInterfacesByNetworkName(List) to get list
only, since Collection type is suiable as well in that
context.

The motivation is to allow other caller that would like
to provide the Collection type to benefit of this method
without need to cast it first to List.

Change-Id: I28e2d22feced3a615393fc3c1cf80ea087888f61
Signed-off-by: Moti Asayag <masayag@redhat.com>