frontend: Allow console settings per user and pool.
This patch introduces the possibility of setting console options for
user per vm pool and storing them in the browser local storage.

These pool settings are then used as a default settings for all vms in
the pool. If user explicitly changes settings for a single vm, this
specific settings will override pool defaults.

Moreover this patch adds some cleanup code in
ConnectAutomaticallyManager.

Change-Id: I0102ad0f44e4fb826c561ac593a428b51622a377
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1042783