engine: Track the time a VM was last stopped
Today, ovirt provides information on how long a VM has been up.  Users
have requested the ability to see how long a VM has been down.  This
could be useful for determining which VMs need to be powered on for
maintenance (eg. package updates).  As a first step, just track this
information and make it available via the REST API.  In the future, it
could be displayed in the UI in the 'Uptime' column or similar.

This patch makes the following specific changes:
- Add new column last_stop_time to the vm_dynamic table and update views
- When a new VM is created, initialize stop time to the current time
- Catch VM state changes in ResourceManager and update last_stop_time
- Add "stop_time" to the REST API and always show it in the VM resource

Signed-off-by: Adam Litke <alitke@redhat.com>
Change-Id: I0f94a56e0c82246a932fc1800a2c9ef8c5205e13
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1042854