webadmin: user session is kept alive
When events are updated we trigger RefreshActiveModelEvent
event, which forcefully update current active model, and
causes a refresh in user session. As far as I understand
this event should only be called when running an action.

The result set from events is no longer empty when no new
events happened, instead it contains the last X event.
Updated the logic that determines if a new event happened
in the frontend.

Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Signed-off-by: Alexander Wels <awels@redhat.com>
Change-Id: Ib99861bfdc41b98ceac0df7103acff5b9074f851