core: findbugs: inconsistent synchronization
AsyncTasksManager is currently the only class with suffers from
inconsistent synchronization. This patch does not solve those issues (in
an attempt to "not rock the boat", but preserve the current situation).
It does, however, move the exclusion of this check out of
exclude-filters-general.xml and fine-grains it to exclude only
AsyncTasksManager in order to avoid these potentially dangerous bugs in
future classes.

Change-Id: Iafbd7c713584ca08b2e74617e289554425ac2b68
Signed-off-by: Allon Mureinik <amureini@redhat.com>