engine: Skip management network removal if unlabeled
The management network can be labeled, although it
is not the recommended approach: Updating the management
network should be updated on any host (i.e. modifying
a management network traits). However, unlabeling a
management network should not attempt to remove it, as
the engine expects any host to be configured with the
management network.

Change-Id: I9bc86f7c2cd4847892fa0fec15244c286abb088e
Signed-off-by: Moti Asayag <masayag@redhat.com>