packaging: services: remove pid when service start fails
the service is running under non privileged user, so it cannot remove
the pid file only clear it.

if service is dead at validation stage its status is printed as:

  dead but pid file exists

to remedy that, if service fails to start we remove the pid if service
start fails and pid file is zero length.

Related-To: https://bugzilla.redhat.com/989382
Change-Id: I8c5ff88aef181dbf4955bd976c176642290e13d0
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>