webadmin: Allow editing labels on a host interface
Added an edit interface dialog, that allows assigning multiple labels
to it. If its labels have changed, the changes that should be applied
to the networks on the interface are considered; if the configuration
is valid then the dialog is redrawn with those changes, else the
status panel explains the error.

Change-Id: I030cbb7bcef1b05f329f67fc8a30ec5dee4b67f9
Signed-off-by: Lior Vernia <lvernia@redhat.com>