5. core: Change LDAP identifer to ExternalId
Currently the LDAP user and group entities are using an UUID as their
identifier because we are assuming that the identifier used by the
external directory is also an UUID (well, we assume that is something
that fits inside the 128 bits of an UUID). This assumption is not
correct for all types directories.  This patch changes the identifier of
these two classes to a plain array of bytes (managed by the ExternalId
class) so that the directory can use whatever it is appropiate for it

Change-Id: Ib464dfc2353158fccc92922893614a35b5c06081
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>