core: replace explicit field names with constants
In VdsUpdateRunTimeInfo we used to have explicit strings for "appList"
and "status", representing the names of those fields in VmDynamic.
This patch changes that - new constants are declared in VmDynamic,
which hold those strings and VdsUpdateRunTimeInfo now uses them.

In addition, places where we used to check the size of collections just
to see if they are empty or not are replaced with calls to "isEmpty"
method - it is better in terms of readability and for some data
structures it is more efficient.

Change-Id: I614fecc4238cfffbf152b1b39e3c7995974188c2
Signed-off-by: Arik Hadas <ahadas@redhat.com>